The framework of librarian has been built
Until now, the functional page entered after logging in as a librarian has been constructed. The librarian can display books, find books and also add new books to the system, the information of the new books will be stored in the system, and they will also be displayed in the additional text file which contains all book information.